it is a big pain that many of the functions in this module use a global accessor.
these edits created versions of some of the functions that do not use the global accessor.  the old version (with global accessor) was then modified to call the one with the explicit accessor.

the 'random_ta_assignment' function was moved to algo/peer_assignment and renamed 'random_assignment'

some functions were rewritten to be more pythonic.